size_t instead of int

+++ 0x05-pointers_arrays_strings/4-print_rev.c
@@ -7,9 +7,9 @@
  *
  * Return: length of string
  */
-int _strlen(char *s)
+size_t _strlen(char *s)
 {
-	int i;
+	size_t i;
 
 	for (i = 0; s[i] != '\0'; i++)
 	{
